Found bug.  The Hitachi code length is 692bits. However the bits of length is uint8_t = 256 + 256 + 180. So we see 180bits before because of overflow.
Change the type of length to int and the type of buffer to char.
The missed code problem is gone.

I also found some code meanings.

+++ IRremoteInt.h
@@ -198,10 +198,9 @@ typedef struct {
   uint8_t rcvstate;          // state machine
   uint8_t blinkflag;         // TRUE to enable blinking of pin 13 on IR processing
   unsigned int timer;     // state timer, counts 50uS ticks.
-  unsigned int rawbuf[RAWBUF]; // raw data
-  uint8_t rawlen;         // counter of entries in rawbuf
-} 
-irparams_t;
+  uint8_t rawbuf[RAWBUF]; // raw data
+  unsigned int rawlen;         // counter of entries in rawbuf
+} irparams_t;
 
 // Defined in IRremote.cpp
 extern volatile irparams_t irparams;
